:recycle: Frontend - Refactor Service classes to individual functions and renamed them to "client".
