--- refarch-frontend/src/api/HealthService.ts
@@ -1,15 +1,0 @@
-import { defaultResponseHandler, getConfig } from "@/api/fetch-utils";
-import HealthState from "@/types/HealthState";
-
-export default class HealthService {
-  static checkHealth(): Promise<HealthState> {
-    return fetch("actuator/health", getConfig())
-      .then((response) => {
-        defaultResponseHandler(response);
-        return response.json();
-      })
-      .catch((err) => {
-        defaultResponseHandler(err);
-      });
-  }
-}
